Update tictactoe
It probably would be way easier to google a tutorial, but I want to try my own way at this.

Checks all around the mouse, if 2 or more directions around mouse are blue, then it is clickable. Also, changed board back into a string, this string can be reassigned and reprinted after every move.